Update App.js
Made changes to how the nav element and it's children are rendered. REMOVED the double A tag that was being built by hovertext and replaced it with a span. 

If this react business is anything like vanilla JS then the inline styling I have added to the links should work...but if not, it won't OBVIOUSLY!